backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.115.12",
    "passlib>=1.7.4",
    "pydantic-settings>=2.8.1",
    "python-jose[cryptography]>=3.4.0",
    "sqlalchemy>=2.0.39",
]
